Port CpScreenSlideout to the slideout module
`Craft.CpScreenSlideout` is now a real TypeScript subclass of the modern
Slideout class, with the screen loading/update flow, tabs, sidebar, error
handling, and delta-aware submit ported intact and the jQuery surface
(`$content`, `$sidebar`, `data('cpScreen')`, …) preserved as public API.
The global stays `compatify()`-wrapped so `Craft.ElementEditorSlideout`'s
legacy `.extend()`/`this.base()` chain keeps dispatching into it.

Also reworks the module README into usage documentation and trims the
code comments down to current behavior and live constraints.
